Fix parsing of the header of lhe files
The 'nom' permutation! macro doesn't handle the case where some of the
elements in the permutation are optional.
This means that lhe files where either the comment or the header block
are optional could not be parsed using this macro.
So it was replaced with the custom permutation_opt! macro.